+++ api/helpers/songs/queue-pending.js
@@ -7,7 +7,11 @@ module.exports = {
     description: 'Queue songs that have been added to the pending list, such as duplicate underwritings.',
 
     inputs: {
-
+        queue: {
+            type: 'boolean',
+            defaultsTo: false,
+            description: 'If true, instead of executing right away, song tracks will be added to a queue where one song is queued every execution of the check CRON. In addition, this helper call will not resolve until all songs are queued.'
+        }
     },
 
     fn: async function (inputs, exits) {
@@ -17,7 +21,7 @@ module.exports = {
         if (Songs.pending.length > 0)
         {
             var maps = Songs.pending.map(async (track, index) => {
-                await sails.helpers.rest.cmd('LoadTrackToTop', track);
+                await sails.helpers.rest.cmd('LoadTrackToTop', track, 10000, inputs.queue);
                 //wait.for.time(1);
                 delete Songs.pending[index];
                 return true;